requirements.txt
streamlit==1.31.1
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
torch==2.1.2
transformers==4.35.2
sentence-transformers==2.2.2
langchain==0.1.4
langchain-community==0.0.10
langchain-huggingface==0.0.5
faiss-cpu==1.7.4
google-genai==1.66.0
google-generativeai==0.8.6
scikit-learn
joblib 